Changed script for manually preparing
